fix(tlds): get tlds when using ioc-extractor version 2.7
This issue occurs due to the ioc-extrator dependency was save this way -> ioc-extrator: "^2.6.3", then the higher ioc-extrator version 2.x will be installed and the 2.7 version has some breaking changes at least this one causing fanger to fail while loading. Maybe you can't reproduce it in the fanger environment because the ioc-extrator version is saved in the package.lock.json or you already have version 2.6 installed.